bot/requirements.txt
slack-bolt==1.18.1
python-dotenv==1.0.0
pycryptodomex==3.20.0
